+++ saleor/graphql/account/types.py
@@ -7,10 +7,12 @@ from graphene_federation import key
 from ...account import models
 from ...checkout.utils import get_user_checkout
 from ...core.exceptions import PermissionDenied
-from ...core.permissions import AccountPermissions, OrderPermissions
+from ...core.permissions import AccountPermissions, AppPermission, OrderPermissions
 from ...core.tracing import traced_resolver
 from ...order import OrderStatus
-from ...order import models as order_models
+from ..account.utils import requestor_has_access
+from ..app.dataloaders import AppByIdLoader
+from ..app.types import App
 from ..checkout.dataloaders import CheckoutByUserAndChannelLoader, CheckoutByUserLoader
 from ..checkout.types import Checkout
 from ..core.connection import CountableDjangoObjectType
@@ -22,7 +24,11 @@ from ..core.utils import from_global_id_or_error, str_to_enum
 from ..decorators import one_of_permissions_required, permission_required
 from ..giftcard.dataloaders import GiftCardsByUserLoader
 from ..meta.types import ObjectWithMetadata
-from ..order.dataloaders import OrdersByUserLoader
+from ..order.dataloaders import (
+    OrderLineByIdLoader,
+    OrdersByUserLoader,
+    SubscriptionsByUserLoader,
+)
 from ..utils import format_permissions_for_display, get_user_or_app_from_context
 from ..wishlist.resolvers import resolve_wishlist_items_from_user
 from .dataloaders import CustomerEventsByUserLoader
@@ -76,12 +82,10 @@ class Address(CountableDjangoObjectType):
         ]
 
     @staticmethod
-    @traced_resolver
     def resolve_country(root: models.Address, _info):
         return CountryDisplay(code=root.country.code, country=root.country.name)
 
     @staticmethod
-    @traced_resolver
     def resolve_is_default_shipping_address(root: models.Address, _info):
         """Look if the address is the default shipping address of the user.
 
@@ -101,7 +105,6 @@ class Address(CountableDjangoObjectType):
         return False
 
     @staticmethod
-    @traced_resolver
     def resolve_is_default_billing_address(root: models.Address, _info):
         """Look if the address is the default billing address of the user.
 
@@ -131,6 +134,7 @@ class CustomerEvent(CountableDjangoObjectType):
     )
     type = CustomerEventsEnum(description="Customer event type.")
     user = graphene.Field(lambda: User, description="User who performed the action.")
+    app = graphene.Field(App, description="App that performed the action.")
     message = graphene.String(description="Content of the event.")
     count = graphene.Int(description="Number of objects concerned by the event.")
     order = graphene.Field(
@@ -147,7 +151,6 @@ class CustomerEvent(CountableDjangoObjectType):
         only_fields = ["id"]
 
     @staticmethod
-    @traced_resolver
     def resolve_user(root: models.CustomerEvent, info):
         user = info.context.user
         if (
@@ -159,6 +162,15 @@ class CustomerEvent(CountableDjangoObjectType):
         raise PermissionDenied()
 
     @staticmethod
+    def resolve_app(root: models.CustomerEvent, info):
+        requestor = get_user_or_app_from_context(info.context)
+        if requestor_has_access(requestor, root.user, AppPermission.MANAGE_APPS):
+            return (
+                AppByIdLoader(info.context).load(root.app_id) if root.app_id else None
+            )
+        raise PermissionDenied()
+
+    @staticmethod
     def resolve_message(root: models.CustomerEvent, _info):
         return root.parameters.get("message", None)
 
@@ -167,15 +179,11 @@ class CustomerEvent(CountableDjangoObjectType):
         return root.parameters.get("count", None)
 
     @staticmethod
-    @traced_resolver
     def resolve_order_line(root: models.CustomerEvent, info):
         if "order_line_pk" in root.parameters:
-            try:
-                qs = order_models.OrderLine.objects
-                order_line_pk = root.parameters["order_line_pk"]
-                return qs.filter(pk=order_line_pk).first()
-            except order_models.OrderLine.DoesNotExist:
-                pass
+            return OrderLineByIdLoader(info.context).load(
+                root.parameters["order_line_pk"]
+            )
         return None
 
 
@@ -276,12 +284,10 @@ class User(CountableDjangoObjectType):
         ]
 
     @staticmethod
-    @traced_resolver
     def resolve_addresses(root: models.User, _info, **_kwargs):
         return root.addresses.annotate_default(root).all()  # type: ignore
 
     @staticmethod
-    @traced_resolver
     def resolve_checkout(root: models.User, _info, **_kwargs):
         return get_user_checkout(root)
 
@@ -309,24 +315,20 @@ class User(CountableDjangoObjectType):
         )
 
     @staticmethod
-    @traced_resolver
     def resolve_gift_cards(root: models.User, info, **_kwargs):
         return GiftCardsByUserLoader(info.context).load(root.id)
 
     @staticmethod
-    @traced_resolver
     def resolve_user_permissions(root: models.User, _info, **_kwargs):
         from .resolvers import resolve_permissions
 
         return resolve_permissions(root)
 
     @staticmethod
-    @traced_resolver
     def resolve_permission_groups(root: models.User, _info, **_kwargs):
         return root.groups.all()
 
     @staticmethod
-    @traced_resolver
     def resolve_editable_groups(root: models.User, _info, **_kwargs):
         return get_groups_which_user_can_manage(root)
 
@@ -341,12 +343,10 @@ class User(CountableDjangoObjectType):
     @one_of_permissions_required(
         [AccountPermissions.MANAGE_USERS, AccountPermissions.MANAGE_STAFF]
     )
-    @traced_resolver
     def resolve_events(root: models.User, info):
         return CustomerEventsByUserLoader(info.context).load(root.id)
 
     @staticmethod
-    @traced_resolver
     def resolve_orders(root: models.User, info, **_kwargs):
         def _resolve_orders(orders):
             requester = get_user_or_app_from_context(info.context)
@@ -357,15 +357,10 @@ class User(CountableDjangoObjectType):
         return OrdersByUserLoader(info.context).load(root.id).then(_resolve_orders)
 
     @staticmethod
-    @traced_resolver
     def resolve_subscriptions(root: models.User, info, **_kwargs):
-        viewer = info.context.user
-        if viewer.has_perm(OrderPermissions.MANAGE_ORDERS):
-            return root.subscriptions.all()
-        return root.subscriptions.active()
+        return SubscriptionsByUserLoader(info.context).load(root.id)
 
     @staticmethod
-    @traced_resolver
     def resolve_avatar(root: models.User, info, size=None, **_kwargs):
         if root.avatar:
             return Image.get_adjusted(
@@ -377,7 +372,6 @@ class User(CountableDjangoObjectType):
             )
 
     @staticmethod
-    @traced_resolver
     def resolve_stored_payment_sources(root: models.User, info, channel=None):
         from .resolvers import resolve_payment_sources
 
@@ -386,7 +380,6 @@ class User(CountableDjangoObjectType):
         raise PermissionDenied()
 
     @staticmethod
-    @traced_resolver
     def resolve_wishlist(root: models.User, info, **_kwargs):
         return resolve_wishlist_items_from_user(root)
 
@@ -397,7 +390,6 @@ class User(CountableDjangoObjectType):
         return get_user_model().objects.get(email=root.email)
 
     @staticmethod
-    @traced_resolver
     def resolve_language_code(root, _info, **_kwargs):
         return LanguageCodeEnum[str_to_enum(root.language_code)]
 
@@ -452,7 +444,6 @@ class StaffNotificationRecipient(CountableDjangoObjectType):
         only_fields = ["user", "active"]
 
     @staticmethod
-    @traced_resolver
     def resolve_user(root: models.StaffNotificationRecipient, info):
         user = info.context.user
         if user == root.user or user.has_perm(AccountPermissions.MANAGE_STAFF):
@@ -460,7 +451,6 @@ class StaffNotificationRecipient(CountableDjangoObjectType):
         raise PermissionDenied()
 
     @staticmethod
-    @traced_resolver
     def resolve_email(root: models.StaffNotificationRecipient, _info):
         return root.get_email()
 
@@ -484,12 +474,10 @@ class Group(CountableDjangoObjectType):
 
     @staticmethod
     @permission_required(AccountPermissions.MANAGE_STAFF)
-    @traced_resolver
     def resolve_users(root: auth_models.Group, _info):
         return root.user_set.all()
 
     @staticmethod
-    @traced_resolver
     def resolve_permissions(root: auth_models.Group, _info):
         permissions = root.permissions.prefetch_related("content_type").order_by(
             "codename"
@@ -497,7 +485,6 @@ class Group(CountableDjangoObjectType):
         return format_permissions_for_display(permissions)
 
     @staticmethod
-    @traced_resolver
     def resolve_user_can_manage(root: auth_models.Group, info):
         user = info.context.user
         return can_user_manage_group(user, root)