+++ saleor/graphql/core/schema.py
@@ -1,6 +1,5 @@
 import graphene
 
-from ...core.tracing import traced_resolver
 from .mutations import FileUpload
 from .types.common import TaxType
 
@@ -10,7 +9,6 @@ class CoreQueries(graphene.ObjectType):
         TaxType, description="List of all tax rates available from tax gateway."
     )
 
-    @traced_resolver
     def resolve_tax_types(self, info):
         manager = info.context.plugins
         return [