+++ saleor/graphql/order/bulk_mutations/orders.py
@@ -29,5 +29,8 @@ class OrderBulkCancel(BaseBulkMutation):
     def bulk_action(cls, info, queryset):
         for order in queryset:
             cancel_order(
-                order=order, user=info.context.user, manager=info.context.plugins
+                order=order,
+                user=info.context.user,
+                app=info.context.app,
+                manager=info.context.plugins,
             )