+++ saleor/order/actions.py
@@ -69,6 +69,7 @@ from .utils import (
 )
 
 if TYPE_CHECKING:
+    from ..app.models import App
     from ..plugins.manager import PluginsManager
     from ..warehouse.models import Warehouse
 
@@ -80,9 +81,13 @@ QuantityType = int
 
 
 def order_created(
-    order: "Order", user: "User", manager: "PluginsManager", from_draft: bool = False
+    order: "Order",
+    user: "User",
+    app: Optional["App"],
+    manager: "PluginsManager",
+    from_draft: bool = False,
 ):
-    events.order_created_event(order=order, user=user, from_draft=from_draft)
+    events.order_created_event(order=order, user=user, app=app, from_draft=from_draft)
     manager.order_created(order)
     payment = order.get_last_payment()
     if payment:
@@ -90,6 +95,7 @@ def order_created(
             order_captured(
                 order=order,
                 user=user,
+                app=app,
                 amount=payment.total,
                 payment=payment,
                 manager=manager,
@@ -98,18 +104,20 @@ def order_created(
             order_authorized(
                 order=order,
                 user=user,
+                app=app,
                 amount=payment.total,
                 payment=payment,
                 manager=manager,
             )
     site_settings = Site.objects.get_current().settings
     if site_settings.automatically_confirm_all_new_orders:
-        order_confirmed(order, user, manager)
+        order_confirmed(order, user, app, manager)
 
 
 def order_confirmed(
     order: "Order",
     user: "User",
+    app: Optional["App"],
     manager: "PluginsManager",
     send_confirmation_email: bool = False,
 ):
@@ -117,16 +125,19 @@ def order_confirmed(
 
     Trigger event, plugin hooks and optionally confirmation email.
     """
-    events.order_confirmed_event(order=order, user=user)
+    events.order_confirmed_event(order=order, user=user, app=app)
     manager.order_confirmed(order)
     if send_confirmation_email:
-        send_order_confirmed(order, user, manager)
+        send_order_confirmed(order, user, app, manager)
 
 
 def handle_fully_paid_order(
-    manager: "PluginsManager", order: "Order", user: Optional["User"] = None
+    manager: "PluginsManager",
+    order: "Order",
+    user: Optional["User"] = None,
+    app: Optional["App"] = None,
 ):
-    events.order_fully_paid_event(order=order, user=user)
+    events.order_fully_paid_event(order=order, user=user, app=app)
     if order.get_customer_email():
         send_payment_confirmation(order, manager)
 
@@ -145,13 +156,18 @@ def handle_fully_paid_order(
 
 
 @traced_atomic_transaction()
-def cancel_order(order: "Order", user: Optional["User"], manager: "PluginsManager"):
+def cancel_order(
+    order: "Order",
+    user: Optional["User"],
+    app: Optional["App"],
+    manager: "PluginsManager",
+):
     """Cancel order.
 
     Release allocation of unfulfilled order items.
     """
 
-    events.order_canceled_event(order=order, user=user)
+    events.order_canceled_event(order=order, user=user, app=app)
 
     deallocate_stock_for_order(order)
     order.status = OrderStatus.CANCELED
@@ -160,40 +176,45 @@ def cancel_order(order: "Order", user: Optional["User"], manager: "PluginsManage
     manager.order_cancelled(order)
     manager.order_updated(order)
 
-    send_order_canceled_confirmation(order, user, manager)
+    send_order_canceled_confirmation(order, user, app, manager)
 
 
 def order_refunded(
     order: "Order",
     user: Optional["User"],
+    app: Optional["App"],
     amount: "Decimal",
     payment: "Payment",
     manager: "PluginsManager",
 ):
     events.payment_refunded_event(
-        order=order, user=user, amount=amount, payment=payment
+        order=order, user=user, app=app, amount=amount, payment=payment
     )
     manager.order_updated(order)
 
-    send_order_refunded_confirmation(order, user, amount, payment.currency, manager)
+    send_order_refunded_confirmation(
+        order, user, app, amount, payment.currency, manager
+    )
 
 
 def order_voided(
     order: "Order",
     user: Optional["User"],
+    app: Optional["App"],
     payment: "Payment",
     manager: "PluginsManager",
 ):
-    events.payment_voided_event(order=order, user=user, payment=payment)
+    events.payment_voided_event(order=order, user=user, app=app, payment=payment)
     manager.order_updated(order)
 
 
 def order_returned(
     order: "Order",
     user: Optional["User"],
+    app: Optional["App"],
     returned_lines: List[Tuple[QuantityType, OrderLine]],
 ):
-    order_returned_event(order=order, user=user, returned_lines=returned_lines)
+    order_returned_event(order=order, user=user, app=app, returned_lines=returned_lines)
     update_order_status(order)
 
 
@@ -201,6 +222,7 @@ def order_returned(
 def order_fulfilled(
     fulfillments: List["Fulfillment"],
     user: "User",
+    app: Optional["App"],
     fulfillment_lines: List["FulfillmentLine"],
     manager: "PluginsManager",
     notify_customer=True,
@@ -208,7 +230,7 @@ def order_fulfilled(
     order = fulfillments[0].order
     update_order_status(order)
     events.fulfillment_fulfilled_items_event(
-        order=order, user=user, fulfillment_lines=fulfillment_lines
+        order=order, user=user, app=app, fulfillment_lines=fulfillment_lines
     )
     manager.order_updated(order)
 
@@ -220,7 +242,9 @@ def order_fulfilled(
 
     if notify_customer:
         for fulfillment in fulfillments:
-            send_fulfillment_confirmation_to_customer(order, fulfillment, user, manager)
+            send_fulfillment_confirmation_to_customer(
+                order, fulfillment, user, app, manager
+            )
 
 
 def order_shipping_updated(order: "Order", manager: "PluginsManager"):
@@ -231,12 +255,13 @@ def order_shipping_updated(order: "Order", manager: "PluginsManager"):
 def order_authorized(
     order: "Order",
     user: Optional["User"],
+    app: Optional["App"],
     amount: "Decimal",
     payment: "Payment",
     manager: "PluginsManager",
 ):
     events.payment_authorized_event(
-        order=order, user=user, amount=amount, payment=payment
+        order=order, user=user, app=app, amount=amount, payment=payment
     )
     manager.order_updated(order)
 
@@ -244,27 +269,30 @@ def order_authorized(
 def order_captured(
     order: "Order",
     user: Optional["User"],
+    app: Optional["App"],
     amount: "Decimal",
     payment: "Payment",
     manager: "PluginsManager",
 ):
     events.payment_captured_event(
-        order=order, user=user, amount=amount, payment=payment
+        order=order, user=user, app=app, amount=amount, payment=payment
     )
     manager.order_updated(order)
     if order.is_fully_paid():
-        handle_fully_paid_order(manager, order, user)
+        handle_fully_paid_order(manager, order, user, app)
 
 
 def fulfillment_tracking_updated(
     fulfillment: "Fulfillment",
     user: "User",
+    app: Optional["App"],
     tracking_number: str,
     manager: "PluginsManager",
 ):
     events.fulfillment_tracking_updated_event(
         order=fulfillment.order,
         user=user,
+        app=app,
         tracking_number=tracking_number,
         fulfillment=fulfillment,
     )
@@ -275,6 +303,7 @@ def fulfillment_tracking_updated(
 def cancel_fulfillment(
     fulfillment: "Fulfillment",
     user: "User",
+    app: Optional["App"],
     warehouse: "Warehouse",
     manager: "PluginsManager",
 ):
@@ -285,11 +314,12 @@ def cancel_fulfillment(
     fulfillment = Fulfillment.objects.select_for_update().get(pk=fulfillment.pk)
     restock_fulfillment_lines(fulfillment, warehouse)
     events.fulfillment_canceled_event(
-        order=fulfillment.order, user=user, fulfillment=fulfillment
+        order=fulfillment.order, user=user, app=app, fulfillment=fulfillment
     )
     events.fulfillment_restocked_items_event(
         order=fulfillment.order,
         user=user,
+        app=app,
         fulfillment=fulfillment,
         warehouse_pk=warehouse.pk,
     )
@@ -303,6 +333,7 @@ def cancel_fulfillment(
 def mark_order_as_paid(
     order: "Order",
     request_user: "User",
+    app: Optional["App"],
     manager: "PluginsManager",
     external_reference: Optional[str] = None,
 ):
@@ -336,7 +367,10 @@ def mark_order_as_paid(
         gateway_response={},
     )
     events.order_manually_marked_as_paid_event(
-        order=order, user=request_user, transaction_reference=external_reference
+        order=order,
+        user=request_user,
+        app=app,
+        transaction_reference=external_reference,
     )
     manager.order_fully_paid(order)
     manager.order_updated(order)
@@ -479,7 +513,7 @@ def automatically_fulfill_digital_lines(order: "Order", manager: "PluginsManager
     fulfill_order_lines(lines_info)
 
     send_fulfillment_confirmation_to_customer(
-        order, fulfillment, user=order.user, manager=manager
+        order, fulfillment, user=order.user, app=None, manager=manager
     )
     update_order_status(order)
 
@@ -573,7 +607,8 @@ def _create_fulfillment_lines(
 
 @traced_atomic_transaction()
 def create_fulfillments(
-    requester: "User",
+    user: "User",
+    app: Optional["App"],
     order: "Order",
     fulfillment_lines_for_warehouses: Dict,
     manager: "PluginsManager",
@@ -585,7 +620,8 @@ def create_fulfillments(
     Next updates Order based on created fulfillments.
 
     Args:
-        requester (User): Requester who trigger this action.
+        user (User): User who trigger this action.
+        app (App): App that trigger the action.
         order (Order): Order to fulfill
         fulfillment_lines_for_warehouses (Dict): Dict with information from which
             system create fulfillments. Example:
@@ -629,7 +665,8 @@ def create_fulfillments(
     transaction.on_commit(
         lambda: order_fulfilled(
             fulfillments,
-            requester,
+            user,
+            app,
             fulfillment_lines,
             manager,
             notify_customer,
@@ -793,7 +830,8 @@ def __get_shipping_refund_amount(
 
 
 def create_refund_fulfillment(
-    requester: Optional["User"],
+    user: Optional["User"],
+    app: Optional["App"],
     order,
     payment,
     order_lines_to_refund: List[OrderLineData],
@@ -816,7 +854,8 @@ def create_refund_fulfillment(
 
     with transaction_with_commit_on_errors():
         total_refund_amount = _process_refund(
-            requester=requester,
+            user=user,
+            app=app,
             order=order,
             payment=payment,
             order_lines_to_refund=order_lines_to_refund,
@@ -881,7 +920,8 @@ def _populate_replace_order_fields(original_order: "Order"):
 
 @traced_atomic_transaction()
 def create_replace_order(
-    requester: Optional["User"],
+    user: Optional[User],
+    app: Optional["App"],
     original_order: "Order",
     order_lines_to_replace: List[OrderLineData],
     fulfillment_lines_to_replace: List[FulfillmentLineData],
@@ -936,7 +976,8 @@ def create_replace_order(
     draft_order_created_from_replace_event(
         draft_order=replace_order,
         original_order=original_order,
-        user=requester,
+        user=user,
+        app=app,
         lines=[(line.quantity, line) for line in lines_to_create],
     )
     return replace_order
@@ -1021,7 +1062,8 @@ def _move_lines_to_replace_fulfillment(
 
 @traced_atomic_transaction()
 def create_return_fulfillment(
-    requester: Optional["User"],
+    user: Optional["User"],
+    app: Optional["App"],
     order: "Order",
     order_lines: List[OrderLineData],
     fulfillment_lines: List[FulfillmentLineData],
@@ -1062,7 +1104,8 @@ def create_return_fulfillment(
         transaction.on_commit(
             lambda: order_returned(
                 order,
-                user=requester,
+                user=user,
+                app=app,
                 returned_lines=returned_lines_list,
             )
         )
@@ -1072,7 +1115,8 @@ def create_return_fulfillment(
 
 @traced_atomic_transaction()
 def process_replace(
-    requester: Optional["User"],
+    user: Optional["User"],
+    app: Optional["App"],
     order: "Order",
     order_lines: List[OrderLineData],
     fulfillment_lines: List[FulfillmentLineData],
@@ -1089,7 +1133,8 @@ def process_replace(
         order=order,
     )
     new_order = create_replace_order(
-        requester=requester,
+        user=user,
+        app=app,
         original_order=order,
         order_lines_to_replace=order_lines,
         fulfillment_lines_to_replace=fulfillment_lines,
@@ -1097,20 +1142,23 @@ def process_replace(
     replaced_lines = [(line.quantity, line) for line in new_order.lines.all()]
     fulfillment_replaced_event(
         order=order,
-        user=requester,
+        user=user,
+        app=app,
         replaced_lines=replaced_lines,
     )
     order_replacement_created(
         original_order=order,
         replace_order=new_order,
-        user=requester,
+        user=user,
+        app=app,
     )
 
     return replace_fulfillment, new_order
 
 
 def create_fulfillments_for_returned_products(
-    requester: Optional["User"],
+    user: Optional["User"],
+    app: Optional["App"],
     order: "Order",
     payment: Optional[Payment],
     order_lines: List[OrderLineData],
@@ -1151,7 +1199,8 @@ def create_fulfillments_for_returned_products(
     with traced_atomic_transaction():
         if refund and payment:
             total_refund_amount = _process_refund(
-                requester=requester,
+                user=user,
+                app=app,
                 order=order,
                 payment=payment,
                 order_lines_to_refund=return_order_lines,
@@ -1167,13 +1216,15 @@ def create_fulfillments_for_returned_products(
         replace_fulfillment, new_order = None, None
         if replace_order_lines or replace_fulfillment_lines:
             replace_fulfillment, new_order = process_replace(
-                requester=requester,
+                user=user,
+                app=app,
                 order=order,
                 order_lines=replace_order_lines,
                 fulfillment_lines=replace_fulfillment_lines,
             )
         return_fulfillment = create_return_fulfillment(
-            requester=requester,
+            user=user,
+            app=app,
             order=order,
             order_lines=return_order_lines,
             fulfillment_lines=return_fulfillment_lines,
@@ -1223,7 +1274,8 @@ def _calculate_refund_amount(
 
 @transaction_with_commit_on_errors()
 def _process_refund(
-    requester: Optional["User"],
+    user: Optional["User"],
+    app: Optional["App"],
     order: "Order",
     payment: Payment,
     order_lines_to_refund: List[OrderLineData],
@@ -1256,21 +1308,23 @@ def _process_refund(
         transaction.on_commit(
             lambda: events.payment_refunded_event(
                 order=order,
-                user=requester,
+                user=user,
+                app=app,
                 amount=amount,  # type: ignore
                 payment=payment,
             )
         )
         transaction.on_commit(
             lambda: send_order_refunded_confirmation(
-                order, requester, amount, payment.currency, manager  # type: ignore
+                order, user, app, amount, payment.currency, manager  # type: ignore
             )
         )
 
     transaction.on_commit(
         lambda: fulfillment_refunded_event(
             order=order,
-            user=requester,
+            user=user,
+            app=app,
             refunded_lines=list(lines_to_refund.values()),
             amount=amount,  # type: ignore
             shipping_costs_included=refund_shipping_costs,